BUG/MEDIUM: compression: Don't forget to update htx_sl and http_msg flags
If the response is compressed, we must update the HTX start-line flags and
the HTTP message flags. It is especially important if there is another
filter enabled. Otherwise, there is no way to know the C-L header was
removed and T-E one was added. Except by looping on headers.

This patch is related to the issue #1660. It must backported as far as 2.0
(for HTX part only).

(cherry picked from commit 39e436e222236c4e6c0efe95c304fe89fd01e111)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>