DOC: remove my name from the config doc
I was surprised to notice that my name was still present as the author
at the top of the config manual. It turns out that this line and a few
other ones in this file remained unchanged since commit 6a06a40501 that
added this doc 15 years ago! It's long been time to get rid of this!

(cherry picked from commit 1401b9cd3ab761de4c9bb72c8367aec298731c6f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>