BUG/MEDIUM: http-ana: Fix memleak in redirect rules with ignore-empty option
A memory leak was introduced when ignore-empty option was added to redirect
rules. If there is no location, when this option is set, the redirection is
aborted and the processing continues. But when this happened, the trash buffer
allocated to format the redirect response was not released.

The bug was introduced by commit bc1223be7 ("MINOR: http-rules: add a new
"ignore-empty" option to redirects.").

This patch should fix the issue #1675. It must be backported to 2.5.

(cherry picked from commit eab175771dc907d3280d173d721267ead67a4dd7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>