BUG/MINOR: map/cli: protect the backref list during "show map" errors
In case of write error in "show map", the backref is detached but
the list wasn't locked when this is done. The risk is very low but
it may happen that two concurrent "show map" one of which would fail
or one "show map" failing while the same entry is being updated could
cause a crash.

This should be backported to all stable versions.

(cherry picked from commit 2edaace575fa1bdfa760ee41d7d7a87efb3586ae)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>