DOC: config: Update doc for PR/PH session states to warn about rewrite failures
When an HTTP header rewrite failure is triggered, and 500-internal-error
response is returned. A "PR" termination state is logged if the error
occurred on the request and "PH" if the error is reported for the response.

The documentation was updated accordingly.

This patch is related to issue #1597.

(cherry picked from commit 24dda9403a7f1ca4a7425ead8da4e3ba4a0cd131)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>