CLEANUP: applet: make appctx_new() initialize the whole appctx
Till now, appctx_new() used to allocate an entry from the pool and
pass it through appctx_init() to initialize a debatable part of it that
didn't correspond anymore to the comments, and fill other fields. It's
hard to say what is fully initialized and what is not.

Let's get rid of that, and always zero the initialization (appctx are
not that big anyway, even with the cache there's no difference in
performance), and initialize what remains. this is cleaner and more
resistant to new field additions.

The appctx_init() function was removed.

(cherry picked from commit 009e42bc59709b032ad7d1c3f5c0310c4bb0b5bf)
[cf: appctx_init() is not removed but recent changes on SSL cli commands require
     a full init of the appctx]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>