MINOR: mux-h2: report a trace event when failing to create a new stream
There are two reasons we can reject the creation of an h2 stream on the
frontend:
  - its creation would violate the MAX_CONCURRENT_STREAMS setting
  - there's no more memory available

And on the backend it's almost the same except that the setting might
have be negotiated after trying to set up the stream.

Let's add traces for such a sitaution so that it's possible to know why
the stream was rejected (currently we only know it was rejected).

It could be nice to backport this to the most recent versions.

(cherry picked from commit e872f75fc453baba9cb07b0383096ab6fcb8fb76)
[wt: adjust ctx: dropped the irrelevant to 2.5 cs_attach_mux() part]
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit aa4a1d39849d9e5b4524202813a6e1154b0c60dd)
Signed-off-by: Willy Tarreau <w@1wt.eu>